+++ src/main/java/rbasamoyai/escalated/EscalatedModsNeoForge.java
@@ -4,7 +4,7 @@ import com.simibubi.create.foundation.utility.CreateLang;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.Block;
-import net.neoforged.fml.ModList;
+import net.neoforged.fml.loading.LoadingModList;
 
 import java.util.Optional;
 import java.util.function.Supplier;
@@ -14,8 +14,12 @@ public enum EscalatedModsNeoForge {
     SABLE;
 
     private final String id;
+    private final boolean isLoaded;
 
-    EscalatedModsNeoForge() { this.id = CreateLang.asId(name()); }
+    EscalatedModsNeoForge() {
+        this.id = CreateLang.asId(name());
+        this.isLoaded = LoadingModList.get().getModFileById(this.id) != null;
+    }
 
     public String id() { return this.id; }
 
@@ -23,7 +27,7 @@ public enum EscalatedModsNeoForge {
 
     public Block getBlock(String id) { return BuiltInRegistries.BLOCK.get(this.resource(id)); }
 
-    public boolean isLoaded() { return ModList.get().isLoaded(this.id); }
+    public boolean isLoaded() { return this.isLoaded; }
 
     public <T> Optional<T> runIfInstalled(Supplier<Supplier<T>> toRun) {
         return this.isLoaded() ? Optional.of(toRun.get().get()) : Optional.empty();